+++ CHANGELOG.md
@@ -14,6 +14,16 @@
 
 # RP2 Change Log
 
+## 0.5.2
+* added Binance.com REST data loader plugin
+* added Binance.com supplemental (CSV) data loader plugin (for autoinvest purchases and ETH to BETH conversions)
+* added CCXT-based pair converter
+* updated documentation
+
+## 0.5.1
+* updated RP2 dependency to latest version, which fixes 2 bugs in HIFO. If you're using HIFO be sure to use RP2 1.10 or better.
+* tweaked documentation
+
 ## 0.5.0
 * add top-level multi-thread support: it's now possible to run data loader plugins in parallel, using the -t option (which selects the number of parallel threads)
 * added Nexo CSV input plugin